# openapi.yaml
openapi: "3.0.3"
info:
  title: CampusXchange API
  version: "1.0.0"
  description: |
    Student peer-to-peer marketplace REST API.

    **Base URL:** `http://localhost:8080`
    **Swagger UI:** `http://localhost:8080/swagger-ui.html`

    ## Authentication
    Protected endpoints require two headers:
    ```
    Authorization: Bearer <access_token>
    X-User-Id: <userId>
    ```
    Both values are returned by `/api/auth/login`, `/api/auth/register`, and `/api/auth/google`.

    ## Real-time Chat (WebSocket)
    Connect via SockJS + STOMP to `http://localhost:8080/ws`.
    - Publish to `/app/chat/send` (include `X-User-Id` as STOMP header)
    - Subscribe to `/topic/chat/{min_uid}_{max_uid}` where the UIDs are sorted ascending
servers:
  - url: http://localhost:8080
    description: Local development
security: []   # overridden per-operation where required
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateProductRequest:
      type: object
      required: [title, description, price, category, condition, college, location]
      properties:
        title:
          type: string
          minLength: 5
          maxLength: 200
        description:
          type: string
          minLength: 20
          maxLength: 5000
        price:
          type: number
          format: decimal
          minimum: 0.01
        originalPrice:
          type: number
          format: decimal
          nullable: true
        category:
          $ref: "#/components/schemas/ProductCategory"
        condition:
          $ref: "#/components/schemas/ProductCondition"
        college:
          type: string
        location:
          type: string
          minLength: 3
          maxLength: 100
        imageUrls:
          type: array
          items:
            type: string
            format: uri
          minItems: 1
          maxItems: 10
          description: Cloudinary URLs — upload directly from frontend before calling this endpoint
        videoUrl:
          type: string
          format: uri
          nullable: true
        modelUrl:
          type: string
          format: uri
          nullable: true
          description: URL to a 3-D model file for the Three.js viewer
    ProductDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        price:
          type: number
          format: decimal
        originalPrice:
          type: number
          format: decimal
          nullable: true
        category:
          $ref: "#/components/schemas/ProductCategory"
        condition:
          $ref: "#/components/schemas/ProductCondition"
        status:
          $ref: "#/components/schemas/ProductStatus"
        sellerId:
          type: integer
          format: int64
        sellerName:
          type: string
        sellerRating:
          type: number
          format: double
          nullable: true
        college:
          type: string
        location:
          type: string
        imageUrls:
          type: array
          items:
            type: string
            format: uri
        videoUrl:
          type: string
          format: uri
          nullable: true
        modelUrl:
          type: string
          format: uri
          nullable: true
        viewCount:
          type: integer
        rating:
          type: number
          format: double
        totalReviews:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # ── Reviews ───────────────────────────────────────────────────────────
    CreateReviewRequest:
      type: object
      required: [productId, sellerId, rating, comment]
      properties:
        productId:
          type: integer
          format: int64
        sellerId:
          type: integer
          format: int64
        rating:
          type: number
          minimum: 1
          maximum: 5
        title:
          type: string
          maxLength: 100
          nullable: true
        comment:
          type: string
          minLength: 10
          maxLength: 2000
        isVerifiedPurchase:
          type: boolean
          default: false
    ReviewDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        productId:
          type: integer
          format: int64
        productTitle:
          type: string
        reviewerId:
          type: integer
          format: int64
        reviewerName:
          type: string
        sellerId:
          type: integer
          format: int64
        sellerName:
          type: string
        rating:
          type: number
        title:
          type: string
          nullable: true
        comment:
          type: string
        isVerifiedPurchase:
          type: boolean
        helpfulCount:
          type: integer
        unhelpfulCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # ── Messages ──────────────────────────────────────────────────────────
...
    ProductCategory:
      type: string
      enum:
        - ELECTRONICS
        - BOOKS
        - FURNITURE
        - CLOTHING
        - SPORTS
        - STATIONERY
        - FOOD
        - TRANSPORT
        - GADGETS
        - NOTES
        - HOSTEL_ESSENTIALS
        - GAMING
        - OTHER
    ProductCondition:
      type: string
      enum: [NEW, LIKE_NEW, GOOD, FAIR, POOR]
    ProductStatus:
      type: string
      enum: [AVAILABLE, SOLD, RESERVED, DELETED]

    # ── Common ────────────────────────────────────────────────────────────
...
    Page:
      type: object
      description: Spring Data pagination wrapper
      properties:
        content:
          type: array
          items: {}
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer
          description: Current page (0-indexed)
        first:
          type: boolean
        last:
          type: boolean
    MessageResponse:
      type: object
      properties:
        message:
          type: string

# ── Paths ──────────────────────────────────────────────────────────────────
paths:
...
  /api/products:
    get:
      tags: [Products]
      summary: List all available products (paginated)
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 0
        - in: query
          name: size
          schema:
            type: integer
            default: 12
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Page"
                  - type: object
                    properties:
                      content:
                        type: array
                        items:
                          $ref: "#/components/schemas/ProductDTO"

    post:
      tags: [Products]
      summary: Create a new product listing
      security:
        - BearerAuth: []
      parameters:
        - in: header
          name: X-User-Id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDTO"
        "400":
          description: Validation error
        "401":
          description: Unauthenticated
  /api/products/{id}:
    get:
      tags: [Products]
      summary: Get a product by ID (also increments view count)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDTO"
        "404":
          description: Product not found

    put:
      tags: [Products]
      summary: Update a product listing (seller only)
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
        - in: header
          name: X-User-Id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDTO"
        "401":
          description: Unauthenticated
        "403":
          description: Not the seller
        "404":
          description: Product not found

    delete:
      tags: [Products]
      summary: Soft-delete a product (seller only)
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
        - in: header
          name: X-User-Id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: Unauthenticated
        "403":
          description: Not the seller
  /api/products/{id}/mark-sold:
    patch:
      tags: [Products]
      summary: Mark a product as sold (seller only)
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
        - in: header
          name: X-User-Id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /api/products/search:
    get:
      tags: [Products]
      summary: Full-text search by title or description
      parameters:
        - in: query
          name: query
          required: true
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 0
        - in: query
          name: size
          schema:
            type: integer
            default: 12
      responses:
        "200":
          description: OK
  /api/products/search/advanced:
    get:
      tags: [Products]
      summary: Filtered search with price range, category, and college
      parameters:
        - in: query
          name: minPrice
          schema:
            type: number
            format: decimal
        - in: query
          name: maxPrice
          schema:
            type: number
            format: decimal
        - in: query
          name: category
          schema:
            $ref: "#/components/schemas/ProductCategory"
        - in: query
          name: college
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 0
        - in: query
          name: size
          schema:
            type: integer
            default: 12
      responses:
        "200":
          description: OK
  /api/products/category/{category}:
    get:
      tags: [Products]
      summary: List products by category
      parameters:
        - in: path
          name: category
          required: true
          schema:
            $ref: "#/components/schemas/ProductCategory"
        - in: query
          name: page
          schema:
            type: integer
            default: 0
        - in: query
          name: size
          schema:
            type: integer
            default: 12
      responses:
        "200":
          description: OK
  /api/products/college/{college}:
    get:
      tags: [Products]
      summary: List products from a specific college
      parameters:
        - in: path
          name: college
          required: true
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 0
        - in: query
          name: size
          schema:
            type: integer
            default: 12
      responses:
        "200":
          description: OK
  /api/products/seller/{sellerId}:
    get:
      tags: [Products]
      summary: List products by a specific seller
      parameters:
        - in: path
          name: sellerId
          required: true
          schema:
            type: integer
            format: int64
        - in: query
          name: page
          schema:
            type: integer
            default: 0
        - in: query
          name: size
          schema:
            type: integer
            default: 12
      responses:
        "200":
          description: OK

  # ── Reviews ───────────────────────────────────────────────────────────
  /api/reviews:
    post:
      tags: [Reviews]
      summary: Create a review for a product / seller
      security:
        - BearerAuth: []
      parameters:
        - in: header
          name: X-User-Id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewDTO"
        "401":
          description: Unauthenticated
  /api/reviews/product/{productId}:
    get:
      tags: [Reviews]
      summary: Get reviews for a product
      parameters:
        - in: path
          name: productId
          required: true
          schema:
            type: integer
            format: int64
        - in: query
          name: page
          schema:
            type: integer
            default: 0
        - in: query
          name: size
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Page"
                  - type: object
                    properties:
                      content:
                        type: array
                        items:
                          $ref: "#/components/schemas/ReviewDTO"
  /api/reviews/seller/{sellerId}:
    get:
      tags: [Reviews]
      summary: Get reviews for a seller
      parameters:
        - in: path
          name: sellerId
          required: true
          schema:
            type: integer
            format: int64
        - in: query
          name: page
          schema:
            type: integer
            default: 0
        - in: query
          name: size
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: OK
  /api/reviews/{id}:
    put:
      tags: [Reviews]
      summary: Update a review (reviewer only)
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
        - in: header
          name: X-User-Id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewDTO"

    delete:
      tags: [Reviews]
      summary: Delete a review (reviewer only)
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
        - in: header
          name: X-User-Id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── Messages ──────────────────────────────────────────────────────────
...