# src/main/java/com/campusxchange/entity/ProductCategory.java
package com.campusxchange.entity;

/**
 * Product Category Enum - Available product categories on CampusXchange
 */
public enum ProductCategory {
    ELECTRONICS,
    BOOKS,
    FURNITURE,
    VEHICLES,
    FASHION,
    ACCESSORIES,
    SPORTS,
    GADGETS,
    STUDY_MATERIALS,
    GAMING,
    HOSTEL_ESSENTIALS,
    COLLEGE_ACCESSORIES,
    STATIONERY,
    NOTES_ACADEMIC_MATERIALS,
    OTHER
}